+++ src/NativeAppstackReactNative.ts
@@ -3,9 +3,10 @@ import { TurboModuleRegistry, NativeModules } from 'react-native';
 
 export interface Spec extends TurboModule {
   configure(apiKey: string, isDebug?: boolean, endpointBaseUrl?: string, logLevel?: number): Promise<boolean>;
-  sendEvent(eventType: string | null, eventName: string | null, revenue?: number | null): Promise<boolean>;
+  sendEvent(eventType: string | null, eventName: string | null, parameters?: Record<string, any> | null): Promise<boolean>;
   enableAppleAdsAttribution(): Promise<boolean>;
   getAppstackId(): Promise<string>;
+  isSdkDisabled(): Promise<boolean>;
 }
 
 // Support both old and new architecture